feat (server): added more checks & logic in restaurant handlers, coordinates query helper

+++ server/src/routes/restaurant-routes.ts
@@ -2,6 +2,7 @@ import { Hono } from "hono";
 import { z } from "zod";
 import authenticationMiddleware from "@/middlewares/authentication";
 import {
+	checkRestaurantOwner,
 	createRestaurant,
 	deleteRestaurant,
 	getAllRestaurants,
@@ -10,11 +11,12 @@ import {
 } from "@/repositories/restaurant-repository";
 import jsonValidator from "@/middlewares/validation";
 import type { ContextWithUser } from "@/types/context";
+import { HTTPException } from "hono/http-exception";
 
 const restaurant = new Hono<ContextWithUser>();
 
 // TODO: only disable authentication middleware for dev purposes
-// restaurant.use("/*", authenticationMiddleware);
+restaurant.use("/*", authenticationMiddleware);
 
 // TODO: move this somewhere else?
 const validationSchema = z.object({
@@ -22,8 +24,10 @@ const validationSchema = z.object({
 	description: z.string(),
 	address: z.string(),
 	category: z.string(),
-	lat: z.number(),
-	lng: z.number(),
+	coordinates: z.object({
+		lat: z.number(),
+		lng: z.number(),
+	}),
 });
 
 restaurant.get("/", async (c) => {
@@ -34,19 +38,33 @@ restaurant.get("/", async (c) => {
 
 restaurant.get("/:id", async (c) => {
 	const id = parseInt(c.req.param("id"));
+
+	if (!id)
+		throw new HTTPException(404, { message: "Record not found, invalid ID" });
+
 	const data = await getRestaurantById(id);
 
 	return c.json({ data });
 });
 
 restaurant.post("/", jsonValidator(validationSchema), async (c) => {
-	const data = await createRestaurant(c.req.valid("json"));
+	const data = await createRestaurant({
+		owner_id: c.var.user.id,
+		...c.req.valid("json"),
+	});
 
 	return c.json({ message: "Restaurant details", data });
 });
 
 restaurant.put("/:id", jsonValidator(validationSchema.partial()), async (c) => {
 	const id = parseInt(c.req.param("id"));
+
+	if (!id)
+		throw new HTTPException(404, { message: "Record not found, invalid ID" });
+
+	const check = await checkRestaurantOwner(id, c.var.user.id);
+	if (!check) throw new HTTPException(401, { message: "Unauthorized" });
+
 	const data = await updateRestaurant(id, c.req.valid("json"));
 
 	return c.json({ message: "Successfully updated restaurant", data });
@@ -54,6 +72,13 @@ restaurant.put("/:id", jsonValidator(validationSchema.partial()), async (c) => {
 
 restaurant.delete("/:id", async (c) => {
 	const id = parseInt(c.req.param("id"));
+
+	if (!id)
+		throw new HTTPException(404, { message: "Record not found, invalid ID" });
+
+	const check = await checkRestaurantOwner(id, c.var.user.id);
+	if (!check) throw new HTTPException(401, { message: "Unauthorized" });
+
 	const data = await deleteRestaurant(id);
 
 	return c.json({ message: "Successfully deleted restaurant", data });
